Moving Zepr class to package
