+++ desktop/src/com/geeselightning/zepr/desktop/DesktopLauncher.java
@@ -2,7 +2,7 @@ package com.geeselightning.zepr.desktop;
 
 import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
 import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.geeselightning.zepr.Zepr;
+import com.geeselightning.zepr.game.Zepr;
 
 public class DesktopLauncher {
 	public static void main (String[] arg) {
